Move all config under one roof
