+++ config/main.js
